Minor tweak to Integer.  It seems on Android it couldn't figure out which sqrt function to call.  Now explicitly cast to float.
git-svn-id: svn://10.0.50.55/puffbomb/trunk@3047 d14347b8-48fd-0310-9e84-97907d434d85

+++ Code/Library/Math/Integer.h
@@ -241,7 +241,7 @@ public:
 	// Square Root //
 	inline const Integer Sqrt() const {
 		if ( !IsZero() )
-			return (int)std::sqrt( Value );
+			return (int)std::sqrt( (float)Value );
 		else
 			return 0;
 	}
